color

+++ matrix.c
@@ -6,7 +6,7 @@
 /*   By: mbico <mbico@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/15 23:03:01 by mbico             #+#    #+#             */
-/*   Updated: 2024/01/17 18:37:44 by mbico            ###   ########.fr       */
+/*   Updated: 2024/01/21 16:03:05 by mbico            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,20 +38,18 @@ void	ft_cmatrix_application(t_vars *vars, int i, int j)
 {
 	int			x;
 	int			y;
-	int			z;
+	double		z;
 	double		**m;
 
 	x = i - vars->maxx / 2;
 	y = j - vars->maxy / 2;
-	z = vars->links[j][i].z / 10 * vars->ampl;
+	z = vars->links[j][i].z / (double)10 * vars->ampl;
 	m = vars->cmatrix.content;
 	vars->links[j][i].x = x * m[0][0] + y * m[0][1] + z * m[0][2] + 512 + vars->viewx;
 	vars->links[j][i].y = x * m[1][0] + y * m[1][1] + z * m[1][2] + 300 + vars->viewy;
 }
 
-
-
-void	ft_matrix_mult_one(t_matrix *m, int nb)
+void	ft_matrix_mult_one(t_matrix *m, double nb)
 {
 	int	i;
 	int	j;
@@ -69,7 +67,7 @@ void	ft_matrix_mult_one(t_matrix *m, int nb)
 	}
 }
 
-void	ft_matrix_div_one(t_matrix *m, int nb)
+void	ft_matrix_div_one(t_matrix *m, double nb)
 {
 	int	i;
 	int	j;
